% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boottest.R
\name{boottest}
\alias{boottest}
\title{Bootstrap Procedure for Testing Statistical Significance of
  Correlation Values}
\usage{
boottest(
  countMat,
  assoMat,
  nboot = 1000,
  measure,
  measurePar,
  parallel = FALSE,
  cores = 4,
  logFile = NULL,
  verbose = TRUE,
  seed = NULL
)
}
\arguments{
\item{countMat}{matrix containing microbiome data (read counts) for which
the correlations are calculated (rows represent samples, columns represent
taxa)}

\item{assoMat}{matrix containing associations that have been estimated for
\code{countMat}}

\item{nboot}{number of bootstrap samples}

\item{measure}{character specifying the method used for computing the
associations between taxa.}

\item{measurePar}{list with parameters passed to the function for computing
associations/dissimilarities. See details for the respective functions.}

\item{parallel}{if \code{TRUE}, the bootstrap iterations are executed
parallel}

\item{cores}{number of CPU cores used for parallelization}

\item{logFile}{character defining a log file, where the number of iteration
is stored. If NULL, no log file is created. wherein the current iteration
numbers are stored}

\item{verbose}{logical; if \code{TRUE}, the iteration numbers are printed to
the R console}

\item{seed}{an optional seed for reproducibility of the results}
}
\value{
\tabular{ll}{ \code{pvals}\tab calculated p-values\cr
  \code{corrMat}\tab estimated correlation matrix}
}
\description{
Statistical significance of correlations between pairs of
  taxonomic units is tested using a bootstrap procedure as proposed by
  \cite{Friedman and Alm (2012)}.
}
\references{
{\insertRef{friedman2012inferring}{NetCoMi}}
}
