% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_sparcc}
\alias{C_sparcc}
\title{Estimate correlations between taxa via SparCC algorithm}
\usage{
C_sparcc(x, exiter, thresh, repEst)
}
\arguments{
\item{x}{matrix with read counts}

\item{exiter}{maximum number of iterations, where in each step the
strongest correlation is excluded and the basis variances and correlations
are re-estimated based on the reduced data}

\item{thresh}{threshold for excluding component pairs in the
iteration steps (pairs are excluded if the magnitude of their correlation
value is upon the threshold)}

\item{repEst}{estimation process is repeated 'repEst' times to
 account for sampling noise; new fraction samples are drawn from Dirichlet
 distribution in each iteration step

 @references{\insertRef{friedman2012inferring}{NetCoMi}}}
}
\description{
Estimate correlations between taxa via SparCC algorithm
}
