% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_association.R
\name{calc_association}
\alias{calc_association}
\title{Compute associations between taxa}
\usage{
calc_association(countMat, measure, measurePar, verbose)
}
\arguments{
\item{countMat}{numeric read count matrix, where rows are samples and columns
are OTUs/taxa.}

\item{measure}{character giving the measure used for estimating associations
or dissimilarities}

\item{measurePar}{optional list with parameters passed to the function for
estimating associations/dissimilarities}

\item{verbose}{if \code{TRUE}, progress messages are returned.}
}
\description{
Computes associations between taxa or distances between subjects
  for a given read count matrix
}
