% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_props.R
\name{calc_props}
\alias{calc_props}
\title{Calculate network properties}
\usage{
calc_props(
  adjaMat,
  dissMat,
  weighted,
  isempty,
  clustMethod,
  clustPar,
  hubPar,
  hubQuant,
  lnormFit,
  connect,
  weightDeg,
  normDeg,
  normBetw,
  normClose,
  normEigen,
  jaccard = FALSE,
  jaccQuant = NULL
)
}
\arguments{
\item{adjaMat}{adjacency matrix}

\item{dissMat}{dissimilarity matrix}

\item{weighted}{indicated whether the network is weighted}

\item{isempty}{indicator whether the network contains any edges.}

\item{clustMethod}{character indicating the clustering algorithm.}

\item{clustPar}{list with parameters passed to the clustering functions.}

\item{hubPar}{character vector with one or more centrality measures used
for identifying hub nodes. Possible values are \code{degree},
\code{betweenness}, \code{closeness}, and \code{eigenvector}.}

\item{hubQuant}{quantile used for determining hub nodes.}

\item{lnormFit}{hubs are nodes with a centrality value above the 95\%
quantile of the fitted log-normal distribution (if \code{lnormFit = TRUE})
or of the empirical distribution of centrality values.}

\item{connect}{logical indicating whether edge and vertex connectivity should
be calculated. Might be disabled to reduce execution time.}

\item{weightDeg}{if \code{TRUE}, the weighted degree is used.}

\item{normDeg, normBetw, normClose, normEigen}{if \code{TRUE}, a normalized
version of the respective centrality values is returned. By default, all
centralities are normalized.}

\item{jaccard}{shall the Jaccard index be calculated?}

\item{jaccQuant}{quantile for the Jaccard index}
}
\description{
Calculates network properterties for a given adjacency matrix
}
