% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cclasso.R
\name{cclasso}
\alias{cclasso}
\title{CCLasso: Correlation inference of Composition data through Lasso method}
\usage{
cclasso(
  x,
  counts = F,
  pseudo = 0.5,
  sig = NULL,
  lams = 10^(seq(0, -8, by = -0.01)),
  K = 3,
  kmax = 5000,
  verbose = TRUE
)
}
\arguments{
\item{x}{numeric matrix (\emph{n}x\emph{p}) with samples in rows and 
OTUs/taxa in columns.}

\item{counts}{logical indicating whether x constains counts or fractions.
Defaults to \code{FALSE} meaning that x contains fractions so that rows
sum up to 1.}

\item{pseudo}{numeric value giving a pseudo count, which is added to all
counts if \code{counts = TRUE}. Default is 0.5.}

\item{sig}{numeric matrix giving an initial covariance matrix. If \code{NULL}
(default), \code{diag(rep(1, p))} is used.}

\item{lams}{numeric vector specifying the tuning parameter sequences. Default
is \code{10^(seq(0, -8, by = -0.01))}.}

\item{K}{numeric value (integer) giving the folds of crossvalidation.
Defaults to 3.}

\item{kmax}{numeric value (integer) specifying the maximum iteration for
augmented lagrangian method. Default is 5000.}

\item{verbose}{logical indicating whether a progress indicator is shown 
(\code{TRUE} by default).}
}
\value{
A list containing the following elements:
\tabular{ll}{
  \code{cov.w}\tab Covariance estimation\cr
  \code{cor.w}\tab Correlation estimation\cr
  \code{lam}\tab Final tuning parameter}
}
\description{
Implementation of the CCLasso approach 
  \cite{(Fang et al., 2015)}, which is published on GitHub 
  \cite{(Fang, 2016)}. The function is extended by a progress message.
}
\references{
\insertRef{fang2015cclasso}{NetCoMi}\cr\cr
  \insertRef{fang2016cclassoGithub}{NetCoMi}
}
\author{
Fang Huaying, Peking University (R code)\cr
  Stefanie Peschel (documentation)
}
