% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{col_to_transp}
\alias{col_to_transp}
\title{Adding transparancy to a color}
\usage{
col_to_transp(col, percent)
}
\arguments{
\item{col}{color vector specified similar to the \code{col} argument in
\code{\link[grDevices]{col2rgb}}}

\item{percent}{numeric between 0 and 100 giving the level of transparency}
}
\description{
Adding transparancy to a color
}
