% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumNormMat.R
\name{cumNormMat}
\alias{cumNormMat}
\title{Cumulative sum scaling factors}
\usage{
cumNormMat(obj, p = cumNormStatFast(obj), sl = 1000)
}
\description{
Copy of the \code{cumNormMat} function from
  metagenomeSeq package.
}
