% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffnet.R
\name{diffnet}
\alias{diffnet}
\title{Constructing Differential Networks for Microbiome Data}
\usage{
diffnet(
  x,
  diffMethod = "permute",
  discordThresh = 0.8,
  fisherTrans = TRUE,
  nPerm = 1000L,
  permPvalsMethod = "pseudo",
  cores = 1L,
  verbose = TRUE,
  logFile = NULL,
  seed = NULL,
  alpha = 0.05,
  adjust = "lfdr",
  lfdrThresh = 0.2,
  trueNullMethod = "convest",
  pvalsVec = NULL,
  assoPerm = NULL
)
}
\arguments{
\item{x}{an object of class \code{microNet} inheriting from a call to
\code{\link{netConstruct}}}

\item{diffMethod}{character string indicating the method used for determining
differential associations. Possible values are \code{"permute"} (default)
for performing permutation tests according to \cite{Gill et al. (2010)},
\code{"discordant"}, which calls \code{\link[discordant]{discordantRun}}
\cite{(Siska and Kechris, 2016)}, and \code{"fisherTest"} for Fisher's
z-test \cite{(Fisher , 1992)}.}

\item{discordThresh}{numeric value in [0,1].
Only used for the discordant method. Specifies
a threshold for the posterior probability that a pair of
taxa is differentially correlated between the groups. Taxa pairs with a
posterior above this threshold are connected in the network. Defaults to
0.8.}

\item{fisherTrans}{logical. If \code{TRUE} (default), Fisher-transformed
correlations are used for permutation tests.}

\item{nPerm}{integer giving the number of permutations for the permutation
tests. Defaults to 1000L.}

\item{permPvalsMethod}{character indicating the method used for determining
p-values for permutation tests. Currently, "pseudo" is the only available
option (see details).}

\item{cores}{integer indicating the number of CPU cores used for
permutation tests. If cores > 1, the tests are performed parallel.
Is limited to the number of available CPU cores determined by
\code{\link[parallel]{detectCores}}. Defaults to 1L (no parallelization).}

\item{verbose}{logical. If \code{TRUE} (default), progress messages are shown.}

\item{logFile}{a character string defining the name of a log file, which is
created when permutation tests are conducted (therein the current iteration
numbers are stored). Defaults to \code{NULL} so that no file is created.}

\item{seed}{integer giving a seed for reproducibility of the results.}

\item{alpha}{numeric value between 0 and 1 giving the significance niveau.
Significantly different correlations are connected in the network. Defaults
to 0.05.}

\item{adjust}{character indicating the method used for multiple testing
adjustment for the tests for differentially correlated pairs of taxa.
Possible values are "lfdr" (default) for local
false discovery rate correction (via \code{\link[fdrtool]{fdrtool}}),
"adaptBH" for the adaptive Benjamini-Hochberg method \cite{(Benjamini and
Hochberg, 2000)}, or one of the methods provided by
\code{\link[stats]{p.adjust}}.}

\item{lfdrThresh}{defines a threshold for the local fdr if "lfdr" is chosen
as method for multiple testing correction. Defaults to 0.2 meaning
that correlations with a corresponding local fdr less than or equal to 0.2
are identified as significant.}

\item{trueNullMethod}{character indicating the method used for estimating the
proportion of true null hypotheses from a vector of p-values. Used for the
adaptive Benjamini-Hochberg method for multiple testing adjustment (chosen
by \code{adjust = "adaptBH"}). Accepts the provided options of the
\code{method} argument of \code{\link[limma]{propTrueNull}}: "convest"
(default), "lfdr", "mean", and "hist". Can alternatively be "farco" for
the "iterative plug-in method" proposed by \cite{Farcomeni (2007)}.}

\item{pvalsVec}{vector with p-values used for permutation tests. Can be used
for performing another method for multiple testing adjustment without
executing the complete permutation process again. See the example.}

\item{assoPerm}{a list with two elements used for the permutation procedure.
Each entry must contain association matrices for \code{"nPerm"}
permutations. This can be either the \code{"assoPerm"} value as part of the
output returned from \code{diffnet} or from \code{\link{netCompare}}. See
the example.}
}
\value{
The function returns an object of class \code{diffnet}. Depending on
  the performed test method, the output contains the following
  elements:\cr\cr
  \strong{Permutation tests:}
  \tabular{ll}{
  \code{assoMat1,assoMat2}\tab matrices with estimated associations\cr
  \code{pvalsVec}\tab vector with p-values\cr
  \code{pAdjustVec}\tab vector with adjusted p-values\cr
  \code{diffMat}\tab adjacency matrix (absolute difference of associations)\cr
  \code{assoPerm}\tab list with two elements containing the association
  matrices for the permuted data for each group (can be passed to
  \code{diffnet} again)}
  \strong{Discordant:}
  \tabular{ll}{
  \code{assoMat1,assoMat2}\tab matrices with estimated correlations\cr
  \code{diffMat}\tab adjacency matrix (absolute difference of correlations)\cr
  \code{classMat}\tab matrix with classes assigned to a taxa pair\cr
  \code{diffProbs}\tab matrix with posterior probabilities that a taxa pair
  is differentially correlated between the groups}
  \strong{Fisher's z-test:}
  \tabular{ll}{
  \code{assoMat1,assoMat2}\tab matrices with estimated correlations\cr
  \code{pvalsVec}\tab vector with p-values\cr
  \code{pAdjustVec}\tab vector with adjusted p-values\cr
  \code{diffMat}\tab adjacency matrix (absolute difference of correlations)}
}
\description{
Constructs a differential network for objects of class
  \code{microNet}. Three methods for identifying differentially associated
  taxa are provided: Fisher's z-test, permutation test, discordant method.
}
\details{
\strong{Permutation procedure:}\cr
  The null hypothesis of these tests is defined as
  \deqn{H_0: a1_ij - a2_ij = 0,} where \eqn{a1_ij} and \eqn{a2_ij} denote the
  association between taxon i and j in group 1 and 2, respectively.\cr
  To generate a sampling distribution of the differences under \eqn{H_0},
  the group labels are randomly reassigned to the samples while the group
  sizes are kept. The associations are then re-estimated for each permuted
  data set. The p-values are calcutated as the proportion of
  "permutation-differences" being larger than the observed difference. A
  pseudocount is added to the numerator and denominator in order to avoid
  zero p-values. The p-values should be adjusted for multiple testing.
}
\examples{
# load data sets from American Gut Project (from SpiecEasi package)
data("amgut1.filt")

# generate a random group vector
set.seed(123456)
group <- sample(1:2, nrow(amgut1.filt), replace = TRUE)

# network construction:
amgut_net <- netConstruct(amgut1.filt, group = group,
                          measure = "pearson",
                          filtTax = "highestVar",
                          filtTaxPar = list(highestVar = 30),
                          zeroMethod = "pseudo", normMethod = "clr")

### differential network
# Fisher's z-test
amgut_diff1 <- diffnet(amgut_net, diffMethod = "fisherTest")

# network contains no differentially correlated taxa:
\dontrun{
  plot(amgut_diff1)
}

# without multiple testing correction (statistically not correct!)
amgut_diff2 <- diffnet(amgut_net, diffMethod = "fisherTest", adjust = "none")
plot(amgut_diff2)

\dontrun{
# Permutation test
amgut_diff3 <- diffnet(amgut_net, diffMethod = "permute", nPerm = 1000L,
                       cores = 6L, adjust = "lfdr")

# use the p-values again (different adjustment method possible), but without
# estimating the associations again
amgut_diff4 <- diffnet(amgut_net, diffMethod = "permute", nPerm = 1000L,
                       adjust = "none", pvalsVec = amgut_diff3$pvalsVec)
plot(amgut_diff4)

# use the permutation associations again (same result as amgut_diff4)
amgut_diff5 <- diffnet(amgut_net, diffMethod = "permute", nPerm = 1000L,
                       adjust = "none", assoPerm = amgut_diff3$assoPerm)
plot(amgut_diff5)
}

}
\references{
\insertRef{benjamini2000adaptive}{NetCoMi} \cr
  \insertRef{discordant2016}{NetCoMi} \cr
  \insertRef{farcomeni2007some}{NetCoMi} \cr
  \insertRef{fisher1992statistical}{NetCoMi} \cr
  \insertRef{gill2010statistical}{NetCoMi}
}
\seealso{
\code{\link{plot.diffnet}}
}
