% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-getVecNames.R
\name{.getVecNames}
\alias{.getVecNames}
\title{Function for Generating Vector Names}
\usage{
.getVecNames(x)
}
\arguments{
\item{x}{symmetric matrix for which the column and row names should be
returned as a vector}
}
\description{
The function generates names for a vector that contains elements
  of the lower triangle of a matrix. The R code has been copied from
  \code{getNames} (not exported function of the \code{discordant} package)
  and changed such that the names are generated by columns (and not by rows
  as in the original function, which produces implausible results).
}
\keyword{internal}
