% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcoda.R
\name{gcoda}
\alias{gcoda}
\title{gCoda: conditional dependence network inference for compositional data}
\usage{
gcoda(
  x,
  counts = F,
  pseudo = 0.5,
  lambda.min.ratio = 1e-04,
  nlambda = 15,
  ebic.gamma = 0.5,
  cores = 1L,
  verbose = TRUE
)
}
\arguments{
\item{x}{numeric matrix (\emph{n}x\emph{p}) with samples in rows and OTUs/taxa in
columns.}

\item{counts}{logical indicating whether x constains counts or fractions.
Defaults to \code{FALSE} meaning that x contains fractions so that rows
sum up to 1.}

\item{pseudo}{numeric value giving a pseudo count, which is added to all
counts if \code{counts = TRUE}. Default is 0.5.}

\item{lambda.min.ratio}{numeric value specifying lambda(max) / lambda(min).
Defaults to 1e-4.}

\item{nlambda}{numberic value (integer) giving the of tuning parameters.
Defaults to 15.}

\item{ebic.gamma}{numeric value specifying the gamma value of EBIC.
Defaults to 0.5.}

\item{cores}{integer indicating the number of CPU cores used for computation. 
Defaults to 1L. For \code{cores} > 1L, \code{\link[foreach]{foreach}} is 
used for parallel execution.}

\item{verbose}{logical indicating whether a progress indicator is shown 
(\code{TRUE} by default).}
}
\value{
A list containing the following elements:
\tabular{ll}{
  \code{lambda}\tab lambda sequence for compuation of EBIC score\cr
  \code{nloglik}\tab negative log likelihood for lambda sequence\cr
  \code{df}\tab number of edges for lambda sequence\cr
  \code{path}\tab sparse pattern for lambda sequence\cr
  \code{icov}\tab inverse covariance matrix for lambda sequence\cr
  \code{ebic.score}\tab EBIC score for lambda sequence\cr
  \code{refit}\tab sparse pattern with best EBIC score\cr
  \code{opt.icov}\tab inverse covariance matrix with best EBIC score\cr
  \code{opt.lambda}\tab lambda with best EBIC score}
}
\description{
A parallelized implementation of the gCoda approach 
  \cite{(Fang et al., 2017)}, published on GitHub \cite{(Fang, 2016)}.
}
\references{
\insertRef{fang2016gcodaGithub}{NetCoMi}\cr\cr
  \insertRef{fang2017gcoda}{NetCoMi}
}
\author{
Fang Huaying, Peking University (R-Code and documentation)\cr
  Stefanie Peschel (Parts of the documentation; Parallelization)
}
