% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installNetCoMiPacks.R
\name{installNetCoMiPacks}
\alias{installNetCoMiPacks}
\title{Install all packages used within NetCoMi}
\usage{
installNetCoMiPacks(onlyMissing = TRUE, lib = NULL, ...)
}
\arguments{
\item{onlyMissing}{logical. If \code{TRUE} (default), 
\code{\link[utils]{installed.packages}} is used to read out the packages
installed in the given library and only missing packages are installed. If
\code{FALSE}, all packages are installed or updated (if already installed).}

\item{lib}{character vector giving the library directories where to install 
missing packages. If \code{NULL}, the first element of 
\code{\link{.libPaths}} is used.}

\item{...}{Additional arguments used by \code{\link[BiocManager]{install}} or
\code{\link[utils]{install.packages}}.}
}
\description{
This function installs the R packages used in NetCoMi not listed
  as dependencies or imports in NetCoMi's description file. These are 
  optional packages only needed in certain network construction settings.\cr
  
  \code{BiocManager::\link[BiocManager]{install}} is used for installation 
  since it installs or updates Bioconductor as well as CRAN packages.\cr
  
  Installed CRAN packages:
  \itemize{
  \item cccd
  \item LaplacesDemon
  \item propr
  \item zCompositions
  }
  
  Installed Bioconductor packages:
  \itemize{
  \item ccrepe
  \item DESeq2
  \item discordant
  \item limma
  \item metagenomeSeq
  }
  
  If not installed via this function, the packages are installed by the 
  respective NetCoMi functions when needed.
}
