% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multAdjust.R
\name{multAdjust}
\alias{multAdjust}
\title{Multiple testing adjustment}
\usage{
multAdjust(
  pvals,
  adjust = "adaptBH",
  trueNullMethod = "convest",
  pTrueNull = NULL,
  verbose = FALSE
)
}
\arguments{
\item{pvals}{numeric vector with p-values}

\item{adjust}{character specifying the method used for adjustment.
Can be \code{"lfdr"}, \code{"adaptBH"}, or one of the methods provided by
\code{\link[stats]{p.adjust}}.}

\item{trueNullMethod}{character indicating the method used for estimating the
proportion of true null hypotheses from a vector of p-values. Used for the
adaptive Benjamini-Hochberg method for multiple testing adjustment (chosen
by \code{adjust = "adaptBH"}). Accepts the provided options of the
\code{method} argument of \code{\link[limma]{propTrueNull}}:
\code{"convest"}(default), \code{"lfdr"}, \code{"mean"}, and \code{"hist"}.
Can alternatively be \code{"farco"} for
the "iterative plug-in method" proposed by \cite{Farcomeni (2007)}.}

\item{pTrueNull}{proportion of true null hypothesis used for the adaptBH 
method. If \code{NULL}, the proportion is computed using the method 
defined via \code{trueNullMethod}.}

\item{verbose}{if \code{TRUE}, progress messages are returned.}
}
\description{
The functions adjusts a vector of p-values for multiple testing
}
\references{
\insertRef{farcomeni2007some}{NetCoMi}
}
