% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCoMi-package.R
\docType{package}
\name{NetCoMi-package}
\alias{NetCoMi}
\alias{NetCoMi-package}
\title{NetCoMi: Network Comparison for Microbial Compositional Data}
\description{
NetCoMi offers functions for constructing, analyzing, and comparing
microbial association networks as well as dissimilarity-based networks for
compositional data. It also includes a function for constructing
differential association networks.\cr\cr
The main functions are \code{\link{netConstruct}}, \code{\link{netAnalyze}},
\code{\link{netCompare}}, and \code{\link{diffnet}}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://netcomi.de}
  \item \url{https://github.com/stefpeschel/NetCoMi}
  \item Report bugs at \url{https://github.com/stefpeschel/NetCoMi/issues}
}

}
\author{
\strong{Maintainer:} Stefanie Peschel \email{stefanie.peschel@mail.de}\cr\cr
  \strong{Acknowledgments:}\cr
  \itemize{
   \item Anne-Laure Boulesteix, Christian L. Müller, and Martin Depner
   (inspiration and theoretical background)
   \item Anastasiia Holovchak (package testing and editing)
  }
}
\keyword{internal}
