% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permtest_diff_asso.R
\name{permtest_diff_asso}
\alias{permtest_diff_asso}
\title{Permutation Tests for Determining Differential Associations}
\usage{
permtest_diff_asso(
  countMat1,
  countMat2,
  assoMat1,
  assoMat2,
  measure,
  measurePar,
  method = c("connect.pairs", "connect.variables", "connect.network"),
  fisherTrans = TRUE,
  pvalsMethod = "pseudo",
  adjust = "lfdr",
  adjust2 = "holm",
  trueNullMethod = "convest",
  alpha = 0.05,
  lfdrThresh = 0.2,
  nPerm = 1000,
  matchDesign = NULL,
  cores = 4,
  verbose = TRUE,
  logFile = "log.txt",
  seed = NULL,
  assoPerm = NULL
)
}
\arguments{
\item{countMat1, countMat2}{matrices containing microbiome data (read counts)
of group 1 and group 2 (rows represent samples and columns taxonomic units,
respectively).}

\item{assoMat1, assoMat2}{association matrices corresponding to the two count
matrices. The associations must have been estimated from the count matrices
\code{countMat1} and \code{countMat2}.}

\item{measure}{a character indicating the association measure that has been
used for determining the association matrices.}

\item{measurePar}{list with parameters passed to the function for association
calculation.}

\item{method}{character vector indicating the tests to be performed. Possible
values are \code{"connect.pairs"} (differentially correlated taxa pairs),
\code{"connect.variables"} (one taxon to all other) and
\code{"connect.network"} (differentially connected networks). By default,
all three tests are conducted.}

\item{fisherTrans}{logical indicating whether the correlation values should
be Fisher-transformed.}

\item{pvalsMethod}{currently only \code{"pseudo"} is available, where 1 is
added to the number of permutations and the permutation test statistics
being more extreme than the observed one in order to avoid zero p-values.}

\item{adjust}{multiple testing adjustment for the tests for differentially
correlated pairs of taxa; possible values are "lfdr" (default) for local
false discovery rate correction (via \code{\link[fdrtool]{fdrtool}}) or one
of the methods provided by \code{\link[stats]{p.adjust}}}

\item{adjust2}{multiple testing adjustment for the tests if a taxa pair is
differentially correlated to all other taxa; possible methods are those
provided by \code{\link[stats]{p.adjust}} (a few hundred tests are
necessary for the local fdr correction)}

\item{trueNullMethod}{character indicating the method used for estimating the
proportion of true null hypotheses from a vector of p-values. Used for the
adaptive Benjamini-Hochberg method for multiple testing adjustment (chosen
by \code{adjust = "adaptBH"}).}

\item{alpha}{significance level}

\item{lfdrThresh}{defines a threshold for the local fdr if "lfdr" is chosen
as method for multiple testing correction; defaults to 0.2, which means
that correlations with a corresponding local fdr less than or equal to 0.2
are identified as significant}

\item{nPerm}{number of permutations}

\item{cores}{number of CPU cores (permutation tests are executed parallel)}

\item{verbose}{if \code{TRUE}, status messages and numbers of SparCC
iterations are printed}

\item{logFile}{character string naming the log file within which the current
iteration number is stored}

\item{seed}{an optional seed for reproducibility of the results}

\item{assoPerm}{list with two elements containing permutation association
matrices for the two groups.}
}
\description{
The function implements procedures to test whether pairs of taxa
  are differentially associated, whether a taxon is differentially associated
  to all other taxa, or whether two networks are differentially associated
  between two groups as proposed  by \cite{Gill et al.(2010)}.
}
\references{
{\insertRef{gill2010statistical}{NetCoMi}}

{\insertRef{knijnenburg2009fewer}{NetCoMi}}
}
\seealso{
\code{\link{diffnet}}
}
