% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diffNet.R
\name{plot.diffnet}
\alias{plot.diffnet}
\title{Plot method for objects of class diffnet}
\usage{
\method{plot}{diffnet}(
  x,
  adjusted = TRUE,
  layout = NULL,
  repulsion = 1,
  labels = NULL,
  shortenLabels = "none",
  labelLength = 6,
  labelPattern = NULL,
  charToRm = NULL,
  labelScale = TRUE,
  labelFont = 1,
  rmSingles = TRUE,
  nodeColor = "gray90",
  nodeTransp = 60,
  borderWidth = 1,
  borderCol = "gray80",
  edgeFilter = "none",
  edgeFilterPar = NULL,
  edgeWidth = 1,
  edgeTransp = 0,
  edgeCol = NULL,
  title = NULL,
  legend = TRUE,
  legendGroupnames = NULL,
  legendTitle = NULL,
  cexNodes = 1,
  cexLabels = 1,
  cexTitle = 1.2,
  cexLegend = 1,
  mar = c(2, 2, 4, 6),
  ...
)
}
\arguments{
\item{x}{object of class \code{diffnet} containing the adjacency matrix 
(absolute differences between associations)}

\item{adjusted}{logical indicating whether the adjacency matrix based on 
adjusted p-values should be used. Defaults to \code{TRUE}. If \code{FALSE}, 
the adjacency matrix is based on non-adjusted p-values. Ignored 
for the discordant method.}

\item{layout}{indicates the layout used for defining node positions. Can be
a character with one of the layouts provided by
\code{\link[qgraph]{qgraph}}: \code{"spring"}(default), \code{"circle"},
or \code{"groups"}. Alternatively, the layouts provided by igraph (see
\code{\link[igraph:layout_]{layout\_}}) are accepted (must be given as
character, e.g. \code{"layout_with_fr"}). Can also be a matrix with row
number equal to the  number of nodes and two columns corresponding to the x
and y coordinate.}

\item{repulsion}{integer specifying repulse radius in the spring layout; for
a value lower than 1, nodes are placed further apart}

\item{labels}{defines the node labels. Can be a character vector with an
entry for each node. If \code{FALSE}, no labels are plotted. Defaults to
the row/column names of the association matrices.}

\item{shortenLabels}{options to shorten node labels. Ignored if node labels
are defined via \code{labels}. Possible options are:
\describe{
\item{\code{"intelligent"}}{Elements of \code{charToRm} are removed, labels
are shortened to length \code{labelLength} and duplicates are removed by
using \code{labelPattern}.}
\item{\code{"simple"}}{Elements of \code{charToRm} are removed and labels
are shortened to length \code{labelLength}.}
\item{\code{"none"}}{Default. Original colnames of adjacency matrices are
used} }}

\item{labelLength}{integer defining the length to which variable names shall
be shortened if \code{shortenLabels} is used. Defaults to 6.}

\item{labelPattern}{vector of three elements, which is only used if argument
\code{shortenLabels} is set to \code{"intelligent"}. If cutting a node label to
length \code{labelLength} leads to duplicates, the label is shortend
according to \code{labelPattern}, where the first entry gives the length of
the first part, the second entry is used a separator, and the third entry
is the length of the second part. Defaults to c(5, "'", 3). If the data
contains, for example, two bacteria "Streptococcus" and "Streptomyces",
they are by default shortened to "Strep'coc" and "Strep'myc".}

\item{charToRm}{vector with characters to remove from node names. Ignored if
labels are given via \code{labels}.}

\item{labelScale}{logical. If \code{TRUE}, node labels are scaled accoring to
node size}

\item{labelFont}{integer defining the font of node labels. Defaults to 1.}

\item{rmSingles}{logical. If \code{TRUE}, unconnected nodes are removed.}

\item{nodeColor}{character or numeric value specifying node colors. Can also
be a vector with a color for each node.}

\item{nodeTransp}{an integer between 0 and 100 indicating the transparency of
node colors. 0 means no transparency, 100 means full transparency. Defaults
to 60.}

\item{borderWidth}{numeric specifying the width of node borders. Defaults to
1.}

\item{borderCol}{character specifying the color of node borders. Defaults to
"gray80"}

\item{edgeFilter}{character indicating whether and how edges should be
filtered. Possible values are \code{"none"} (all edges are shown) and
\code{"highestDiff"} (the first x edges with highest absolute difference
are shown). x is defined by \code{edgeFilterPar}.}

\item{edgeFilterPar}{numeric value specifying the "x" in \code{edgeFilter}.}

\item{edgeWidth}{numeric specifying the edge width. See argument
\code{"edge.width"} of \code{\link[qgraph]{qgraph}}.}

\item{edgeTransp}{an integer between 0 and 100 indicating the transparency of
edge colors. 0 means no transparency (default), 100 means full transparency.}

\item{edgeCol}{character vector specifying the edge colors. Must be of length
6 for the discordant method (default: c("hotpink", "aquamarine", "red",
"orange", "green", "blue")) and of lengths 9 for permutation tests and
Fisher's z-test (default: c("chartreuse2", "chartreuse4", "cyan",
"magenta", "orange", "red", "blue", "black", "purple")).}

\item{title}{optional character string for the main title.}

\item{legend}{logical. If \code{TRUE}, a legend is plotted.}

\item{legendGroupnames}{a vector with two elements giving the group names
shown in the legend.}

\item{legendTitle}{character specifying the legend title.}

\item{cexNodes}{numeric scaling node sizes. Defaults to 1.}

\item{cexLabels}{numeric scaling node labels. Defaults to 1.}

\item{cexTitle}{numeric scaling the title. Defaults to 1.2.}

\item{cexLegend}{numeric scaling the legend size. Defaults to 1.}

\item{mar}{a numeric vector of the form c(bottom, left, top, right) defining
the plot margins. Works similar to the \code{mar} argument in
\code{\link[graphics]{par}}. Defaults to c(2,2,4,6).}

\item{...}{further arguments being passed to \code{\link[qgraph]{qgraph}}.}
}
\description{
Plot method for objects of class \code{diffnet} inheriting from a call to
\code{\link{diffnet}}.
}
\seealso{
\code{\link{diffnet}}, \code{\link{netConstruct}}
}
