% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGCD.R
\name{print.GCD}
\alias{print.GCD}
\title{Print method for GCD objects}
\usage{
\method{print}{GCD}(x, ...)
}
\arguments{
\item{x}{object of class \code{GCD} (returned by to \code{\link{calcGCD}}).}

\item{...}{not used.}
}
\description{
Print method for GCD objects
}
