% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparcc_cpp.R
\name{sparcc_cpp}
\alias{sparcc_cpp}
\title{Estimate Correlations for Microbiome Data}
\usage{
sparcc_cpp(countData, exiter = 20, thresh = 0.1, repEst = 100, seed = NULL)
}
\arguments{
\item{countData}{matrix containing microbiome data (read counts), where rows
represent samples and columns the OTUs/taxa}

\item{exiter}{numeric; maximum number of iterations, where in each step the
strongest correlation is excluded and the basis variances and correlations
are re-estimated based on the reduced data}

\item{thresh}{numeric; threshold for excluding component pairs in the
iteration steps (pairs are excluded if the magnitude of their correlation
value is upon the threshold)}

\item{repEst}{numeric; estimation process is repeated 'repEst' times to
account for sampling noise; new fraction samples are drawn from Dirichlet
distribution in each iteration step}

\item{seed}{an optional seed for reproducibility of the results}
}
\value{
\tabular{ll}{ \code{CORR}\tab matrix with estimated correlations\cr
  \code{COV}\tab estimated covariance matrix\cr \code{VBASIS}\tab vector with
  basis variances}
}
\description{
Function for estimating correlations between OTU pairs based on
  the method proposed by \cite{Friedman and Alm (2012)}.
}
\details{
The function is written in C++ for a considerably increased runtime.
  The structure of the R code is adopted from r-sparcc package on GitHub.
}
\references{
\insertRef{friedman2012inferring}{NetCoMi}\cr
  \insertRef{sparccRimplement}{NetCoMi}
}
