% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_functions.R
\name{springLayout}
\alias{springLayout}
\title{Generate a Spring Embedded Layout}
\usage{
springLayout(adjaMat, repulse.rad = NULL)
}
\arguments{
\item{adjaMat}{adjacency matrix}

\item{repulse.rad}{positive numeric value indicating the strength of
repulsive forces in the "spring" layout. Nodes are placed closer together
for smaller values and further apart for higher values.}
}
\description{
The Fruchterman Reingold layout is used to determine node
  positions for a given adjacency matrix. R code from
  \code{\link[qgraph]{qgraph}} is used in this function.
}
