% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netAnalyze_methods.R
\name{summary.microNetProps}
\alias{summary.microNetProps}
\alias{print.summary.microNetProps}
\title{Summary Method for Objects of Class microNetProps}
\usage{
\method{summary}{microNetProps}(
  object,
  groupNames = NULL,
  showCompSize = TRUE,
  showGlobal = TRUE,
  showGlobalLCC = TRUE,
  showCluster = TRUE,
  clusterLCC = FALSE,
  showHubs = TRUE,
  showCentr = "all",
  numbNodes = NULL,
  digits = 5L,
  ...
)

\method{print}{summary.microNetProps}(x, ...)
}
\arguments{
\item{object}{object of class \code{microNetProps} (returned by 
\code{\link{netAnalyze}}).}

\item{groupNames}{character vector with two elements giving the group names
corresponding to the two networks. If \code{NULL}, the names are adopted
from \code{object}. Ignored if \code{object} contains a single network.}

\item{showCompSize}{logical. If \code{TRUE}, the component sizes are printed.}

\item{showGlobal}{logical. If \code{TRUE}, global network properties for the 
whole network are printed.}

\item{showGlobalLCC}{logical. If \code{TRUE}, global network properties for 
the largest connected component are printed. If the network is connected 
(number of components is 1) the global properties are printed only once 
(if one of the arguments \code{showGlobal} and \code{showGlobalLCC}) is 
\code{TRUE}.}

\item{showCluster}{logical. If \code{TRUE}, the cluster(s) are printed.}

\item{clusterLCC}{logical. If \code{TRUE}, clusters are printed only for the
largest connected component. Defaults to \code{FALSE} (whole network).}

\item{showHubs}{logical. If \code{TRUE}, the detected hubs are printed.}

\item{showCentr}{character vector indicating for which centrality measures
the results shall be printed. Possible values are "all", "degree",
"betweenness", "closeness", "eigenvector" and "none".}

\item{numbNodes}{integer indicating for how many nodes the centrality
values shall be printed. Defaults to 10L for a single network and 5L for 
two networks. Thus, in case of a single network, the first 10 
nodes with highest centrality value of the specific centrality measure
are shown. If \code{object} contains two networks, for each centrality 
measure a splitted matrix is shown where the upper part contains the 
highest values of the first group, and the lower part the highest values of 
the second group.}

\item{digits}{integer giving the number of decimal places to which the 
results are rounded. Defaults to 5L.}

\item{...}{not used.}

\item{x}{object of class \code{summary.microNetProps} (returned by 
to \code{\link{summary.microNetProps}}).}
}
\description{
The main results returned by \code{\link{netAnalyze}} are
  printed in a well-arranged format.
}
\seealso{
\code{\link{netConstruct}}, \code{\link{netAnalyze}}
}
