% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testGCM.R
\name{testGCM}
\alias{testGCM}
\title{Test GCM(s) for statistical significance}
\usage{
testGCM(
  obj1,
  obj2 = NULL,
  adjust = "adaptBH",
  lfdrThresh = 0.2,
  trueNullMethod = "convest",
  alpha = 0.05,
  verbose = TRUE
)
}
\arguments{
\item{obj1}{object of class \code{GCM} or \code{GCD} returned by
\code{\link{calcGCM}} or \code{\link{calcGCD}}. See details.}

\item{obj2}{optional object of class \code{GCM} returned by 
\code{\link{calcGCM}}. See details.}

\item{adjust}{character indicating the method used for multiple testing
adjustment.
Possible values are "lfdr" (default) for local
false discovery rate correction (via \code{\link[fdrtool]{fdrtool}}),
"adaptBH" for the adaptive Benjamini-Hochberg method \cite{(Benjamini and
Hochberg, 2000)}, or one of the methods provided by
\code{\link[stats]{p.adjust}}.}

\item{lfdrThresh}{defines a threshold for the local fdr if "lfdr" is chosen
as method for multiple testing correction. Defaults to 0.2 meaning
that differences with a corresponding local fdr less than or equal to 0.2
are identified as significant.}

\item{trueNullMethod}{character indicating the method used for estimating the
proportion of true null hypotheses from a vector of p-values. Used for the
adaptive Benjamini-Hochberg method for multiple testing adjustment (chosen
by \code{adjust = "adaptBH"}). Accepts the provided options of the
\code{method} argument of \code{\link[limma]{propTrueNull}}: "convest"
(default), "lfdr", "mean", and "hist". Can alternatively be "farco" for
the "iterative plug-in method" proposed by \cite{Farcomeni (2007)}.}

\item{alpha}{numeric value between 0 and 1 giving the desired significance 
level.}

\item{verbose}{logical. If \code{TRUE} (default), progress messages 
are printed.}
}
\value{
A list with the following elements:
  \tabular{ll}{
  \code{gcm1}\tab Graphlet Correlatoin Matrix GCM1\cr
  \code{pvals1}\tab Matrix with p-values (H0: gc1_ij = 0)\cr
  \code{padjust1}\tab Matrix with adjusted p-values\cr
  }
  \cr
  Additional elements if two GCMs are given:
  \tabular{ll}{
  \code{gcm2}\tab Graphlet Correlatoin Matrix GCM2\cr
  \code{pvals2}\tab Matrix with p-values (H0: gc2_ij = 0)\cr
  \code{padjust2}\tab Matrix with adjusted p-values\cr
  \code{diff}\tab Matrix with differences between graphlet correlations 
  (GCM1 - GCM2)\cr
  \code{absDiff}\tab Matrix with absolute differences between graphlet 
  correlations (|GCM1 - GCM2|)\cr
  \code{pvalsDiff}\tab Matrix with p-values (H0: |gc1_ij - gc2_ij| = 0)\cr
  \code{pAdjustDiff}\tab Matrix with adjusted p-values\cr
  \code{sigDiff}\tab Same as \code{diff}, but non-significant differences 
  are set to zero.\cr
  \code{sigAbsDiff}\tab Same as \code{absDiff}, but non-significant 
  values are set to zero.
  }
}
\description{
The function tests whether graphlet correlations (entries of 
  the GCM) are significantly different from zero.\cr\cr
  If two GCMs are given, the graphlet correlations of the two networks are 
  tested for being significantly different, i.e., Fishers z-test 
  is performed to test if the absolute differences between graphlet 
  correlations are significantly different from zero.
}
\details{
By applying Student's t-test to the Fisher-transformed correlations, 
  all entries of the GCM(s) are tested for being 
  significantly different from zero:\cr\cr
  H0:  gc_ij = 0  vs.  H1: gc_ij != 0,\cr\cr
  with gc_ij being the graphlet correlations.\cr\cr
  
  If both GCMs are given or \code{obj1} is of class \code{GCD}, the absolute
  differences between graphlet correlations are tested for being different 
  from zero using Fisher's z-test. The hypotheses are:\cr\cr
  H0: |d_ij| = 0 vs. H1: |d_ij| > 0,\cr\cr
  where d_ij = gc1_ij - gc2_ij
}
\examples{
# See help page of calcGCD()
?calcGCD

}
