% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_props.R
\name{calc_props}
\alias{calc_props}
\title{Calculate network properties}
\usage{
calc_props(
  adjaMat,
  dissMat,
  assoMat,
  avDissIgnoreInf,
  sPathNorm,
  sPathAlgo,
  connectivity,
  normNatConnect,
  weighted,
  isempty,
  clustMethod,
  clustPar,
  weightClustCoef,
  hubPar,
  hubQuant,
  lnormFit,
  weightDeg,
  normDeg,
  normBetw,
  normClose,
  normEigen,
  centrLCC,
  jaccard = FALSE,
  jaccQuant = NULL,
  verbose = 0
)
}
\arguments{
\item{adjaMat}{adjacency matrix}

\item{dissMat}{dissimilarity matrix}

\item{assoMat}{association matrix}

\item{avDissIgnoreInf}{logical indicating whether to ignore infinities when 
calculating the average dissimilarity. If \code{FALSE}, infinity values 
are set to 1.}

\item{sPathNorm}{logical. If TRUE, shortest paths are normalized by 
average dissimilarity.}

\item{sPathAlgo}{character indicating the algorithm used for shortest path
calculation.}

\item{connectivity}{logical indicating whether edge and vertex connectivity 
should be calculated. Might be disabled to reduce execution time.}

\item{normNatConnect}{logical indicating whether natural connectivity should
be normalized.}

\item{weighted}{indicated whether the network is weighted}

\item{isempty}{indicator whether the network contains any edges.}

\item{clustMethod}{character indicating the clustering algorithm.}

\item{clustPar}{list with parameters passed to the clustering functions.}

\item{weightClustCoef}{logical. If \code{TRUE}, the weighted clustering
coefficient is used.}

\item{hubPar}{character vector with one or more centrality measures used
for identifying hub nodes. Possible values are \code{degree},
\code{betweenness}, \code{closeness}, and \code{eigenvector}.}

\item{hubQuant}{quantile used for determining hub nodes.}

\item{lnormFit}{hubs are nodes with a centrality value above the 95\%
quantile of the fitted log-normal distribution (if \code{lnormFit = TRUE})
or of the empirical distribution of centrality values.}

\item{weightDeg}{if \code{TRUE}, the weighted degree is used.}

\item{normDeg, normBetw, normClose, normEigen}{if \code{TRUE}, a normalized
version of the respective centrality values is returned. By default, all
centralities are normalized.}

\item{centrLCC}{logical indicating whether centralities should only be 
computed for the largest connected component (LCC). If \code{TRUE} 
(default), centrality values of disconnected components are zero.}

\item{jaccard}{shall the Jaccard index be calculated?}

\item{jaccQuant}{quantile for the Jaccard index}

\item{verbose}{integer indicating the level of verbosity. Possible values:
\code{"0"}: no messages, \code{"1"}: only important messages,
\code{"2"}(default): all progress messages are shown. Can also be logical.}
}
\description{
Calculates network properterties for a given adjacency matrix
}
