% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffnet.R
\name{diffnet}
\alias{diffnet}
\title{Constructing Differential Networks for Microbiome Data}
\usage{
diffnet(
  x,
  diffMethod = "permute",
  discordThresh = 0.8,
  fisherTrans = TRUE,
  nPerm = 1000L,
  permPvalsMethod = "pseudo",
  cores = 1L,
  verbose = TRUE,
  logFile = NULL,
  seed = NULL,
  alpha = 0.05,
  adjust = "lfdr",
  lfdrThresh = 0.2,
  trueNullMethod = "convest",
  pvalsVec = NULL,
  fileLoadAssoPerm = NULL,
  fileLoadCountsPerm = NULL,
  storeAssoPerm = FALSE,
  fileStoreAssoPerm = "assoPerm",
  storeCountsPerm = FALSE,
  fileStoreCountsPerm = c("countsPerm1", "countsPerm2"),
  assoPerm = NULL
)
}
\arguments{
\item{x}{an object of class \code{microNet} inheriting from a call to
\code{\link{netConstruct}}}

\item{diffMethod}{character string indicating the method used for determining
differential associations. Possible values are \code{"permute"} (default)
for performing permutation tests according to \cite{Gill et al. (2010)},
\code{"discordant"}, which calls \code{\link[discordant]{discordantRun}}
\cite{(Siska and Kechris, 2016)}, and \code{"fisherTest"} for Fisher's
z-test \cite{(Fisher , 1992)}.}

\item{discordThresh}{numeric value in [0,1].
Only used for the discordant method. Specifies
a threshold for the posterior probability that a pair of
taxa is differentially correlated between the groups. Taxa pairs with a
posterior above this threshold are connected in the network. Defaults to
0.8.}

\item{fisherTrans}{logical. If \code{TRUE} (default), Fisher-transformed
correlations are used for permutation tests.}

\item{nPerm}{integer giving the number of permutations for the permutation
tests. Defaults to 1000L.}

\item{permPvalsMethod}{character indicating the method used for determining
p-values for permutation tests. Currently, "pseudo" is the only available
option (see details).}

\item{cores}{integer indicating the number of CPU cores used for
permutation tests. If cores > 1, the tests are performed parallel.
Is limited to the number of available CPU cores determined by
\code{\link[parallel]{detectCores}}. Defaults to 1L (no parallelization).}

\item{verbose}{logical. If \code{TRUE} (default), progress messages are shown.}

\item{logFile}{a character string defining the name of a log file, which is
created when permutation tests are conducted (therein the current iteration
numbers are stored). Defaults to \code{NULL} so that no file is created.}

\item{seed}{integer giving a seed for reproducibility of the results.}

\item{alpha}{numeric value between 0 and 1 giving the significance level.
Significantly different correlations are connected in the network. Defaults
to 0.05.}

\item{adjust}{character indicating the method used for multiple testing
adjustment for the tests for differentially correlated pairs of taxa.
Possible values are "lfdr" (default) for local
false discovery rate correction (via \code{\link[fdrtool]{fdrtool}}),
"adaptBH" for the adaptive Benjamini-Hochberg method \cite{(Benjamini and
Hochberg, 2000)}, or one of the methods provided by
\code{\link[stats]{p.adjust}}.}

\item{lfdrThresh}{defines a threshold for the local fdr if "lfdr" is chosen
as method for multiple testing correction. Defaults to 0.2 meaning
that correlations with a corresponding local fdr less than or equal to 0.2
are identified as significant.}

\item{trueNullMethod}{character indicating the method used for estimating the
proportion of true null hypotheses from a vector of p-values. Used for the
adaptive Benjamini-Hochberg method for multiple testing adjustment (chosen
by \code{adjust = "adaptBH"}). Accepts the provided options of the
\code{method} argument of \code{\link[limma]{propTrueNull}}: "convest"
(default), "lfdr", "mean", and "hist". Can alternatively be "farco" for
the "iterative plug-in method" proposed by \cite{Farcomeni (2007)}.}

\item{pvalsVec}{vector with p-values used for permutation tests. Can be used
for performing another method for multiple testing adjustment without
executing the complete permutation process again. See the example.}

\item{fileLoadAssoPerm}{character giving the name (without extension) 
or path of the file storing the "permuted" association/dissimilarity 
matrices that have been exported by setting \code{storeAssoPerm} to 
\code{TRUE}. Only used for permutation tests. Set to \code{NULL} if no 
existing associations should be used.}

\item{fileLoadCountsPerm}{character giving the name (without extension) 
or path of the file storing the "permuted" count matrices that have been 
exported by setting \code{storeCountsPerm} to \code{TRUE}. 
Only used for permutation tests, and if \code{fileLoadAssoPerm = NULL}. 
Set to \code{NULL} if no existing count matrices should be used.}

\item{storeAssoPerm}{logical indicating whether the association (or
dissimilarity) matrices for the permuted data should be stored in a file.
The filename is given via \code{fileStoreAssoPerm}. If \code{TRUE}, 
the computed "permutation" association/dissimilarity matrices can be reused
via \code{fileLoadAssoPerm} to save runtime. Defaults to \code{FALSE}.}

\item{fileStoreAssoPerm}{character giving the file name to store a matrix
containing a matrix with associations/dissimilarities for the permuted 
data. Can also be a path.}

\item{storeCountsPerm}{logical indicating whether the permuted count matrices
should be stored in an external file. Defaults to \code{FALSE}.}

\item{fileStoreCountsPerm}{character vector with two elements giving the 
names of two files storing the permuted count matrices belonging to the 
two groups.}

\item{assoPerm}{only needed for output generated with NetCoMi v1.0.1! A 
list with two elements used for the permutation procedure.
Each entry must contain association matrices for \code{"nPerm"}
permutations. This can be either the \code{"assoPerm"} value as part of the
output returned from \code{diffnet} or from \code{\link{netCompare}}. See
the example.}
}
\value{
The function returns an object of class \code{diffnet}. Depending on
  the performed test method, the output contains the following
  elements:\cr\cr
  \strong{Permutation tests:}
  \tabular{ll}{
  \code{diffMat}\tab matrix with absolute differences of associations that are
  significantly different from zero; optional adjacency matrix\cr
  \code{diffAdjustMat}\tab matrix with absolute differences of associations 
  that are significantly different from zero (after multiple testing 
  correction); optional adjacency matrix\cr
  \code{pvalsVec}\tab vector with p-values\cr
  \code{pAdjustVec}\tab vector with adjusted p-values\cr
  \code{pvalsMat}\tab matrix with p-values\cr
  \code{pAdjustMat}\tab matrix with adjusted p-values\cr
  \code{testStatData}\tab vector with test statistics (absolute differences 
  of associations) for the original data\cr
  \code{testStatPerm}\tab matrix with test statistics (absolute differences 
  of associations) for the permuted data\cr
  \code{assoMat1,assoMat2}\tab matrices with estimated associations (of the
  original data)}
  \strong{Discordant:}
  \tabular{ll}{
  \code{assoMat1,assoMat2}\tab matrices with estimated correlations\cr
  \code{diffMat}\tab adjacency matrix (absolute difference of correlations)\cr
  \code{classMat}\tab matrix with classes assigned to a taxa pair\cr
  \code{diffProbs}\tab matrix with posterior probabilities that a taxa pair
  is differentially correlated between the groups}
  \strong{Fisher's z-test:}
  \tabular{ll}{
  \code{diffMat}\tab matrix with absolute differences of associations that are
  significantly different from zero; optional adjacency matrix\cr
  \code{diffAdjustMat}\tab matrix with absolute differences of associations 
  that are significantly different from zero (after multiple testing 
  correction); optional adjacency matrix\cr
  \code{pvalsVec}\tab vector with p-values\cr
  \code{pAdjustVec}\tab vector with adjusted p-values\cr
  \code{pvalsMat}\tab matrix with p-values\cr
  \code{pAdjustMat}\tab matrix with adjusted p-values\cr
  \code{assoMat1,assoMat2}\tab matrices with estimated associations (of the
  original data)}
}
\description{
Constructs a differential network for objects of class
  \code{microNet}. Three methods for identifying differentially associated
  taxa are provided: Fisher's z-test, permutation test, discordant method.
}
\details{
\strong{Permutation procedure:}\cr
  The null hypothesis of these tests is defined as
  \deqn{H_0: a1_ij - a2_ij = 0,} where \eqn{a1_ij} and \eqn{a2_ij} denote the
  association between taxon i and j in group 1 and 2, respectively.\cr
  To generate a sampling distribution of the differences under \eqn{H_0},
  the group labels are randomly reassigned to the samples while the group
  sizes are kept. The associations are then re-estimated for each permuted
  data set. The p-values are calculated as the proportion of
  "permutation-differences" being larger than the observed difference. A
  pseudo-count is added to the numerator and denominator in order to avoid
  zero p-values. The p-values should be adjusted for multiple testing.
}
\examples{
# Load data sets from American Gut Project (from SpiecEasi package)
data("amgut1.filt")

# Generate a random group vector
set.seed(123456)
group <- sample(1:2, nrow(amgut1.filt), replace = TRUE)

# Network construction:
amgut_net <- netConstruct(amgut1.filt, group = group,
                          measure = "pearson",
                          filtTax = "highestVar",
                          filtTaxPar = list(highestVar = 30),
                          zeroMethod = "pseudo", normMethod = "clr")

#---------------------
# Differential network

# Fisher's z-test
amgut_diff1 <- diffnet(amgut_net, diffMethod = "fisherTest")

# Network contains no differentially correlated taxa:
\dontrun{
  plot(amgut_diff1)
}

# Without multiple testing correction (statistically not correct!)
amgut_diff2 <- diffnet(amgut_net, diffMethod = "fisherTest", adjust = "none")
plot(amgut_diff2)

\dontrun{
  # Permutation test (permutation matrices are stored)
  amgut_diff3 <- diffnet(amgut_net, diffMethod = "permute", nPerm = 1000L,
                         cores = 4L, adjust = "lfdr",
                         storeCountsPerm = TRUE, 
                         fileStoreCountsPerm = c("countsPerm1", "countsPerm2"),
                         storeAssoPerm = TRUE,
                         fileStoreAssoPerm = "assoPerm",
                         seed = 123456)
  
  # Use the p-values again (different adjustment method possible), but without
  # re-estimating the associations
  amgut_diff4 <- diffnet(amgut_net, diffMethod = "permute", nPerm = 1000L,
                         adjust = "none", pvalsVec = amgut_diff3$pvalsVec)
  x11()
  plot(amgut_diff4)
  
  # Use the permutation associations again (same result as amgut_diff4)
  amgut_diff5 <- diffnet(amgut_net, diffMethod = "permute", nPerm = 1000L,
                         adjust = "none", 
                         fileLoadAssoPerm = "assoPerm")
  x11()
  plot(amgut_diff5)
  
  # Use the permuted count matrices again (same result as amgut_diff4)
  amgut_diff6 <- diffnet(amgut_net, diffMethod = "permute", nPerm = 1000L,
                         adjust = "none", 
                         fileLoadCountsPerm = c("countsPerm1", "countsPerm2"),
                         seed = 123456)
  x11()
  plot(amgut_diff6)
}

}
\references{
\insertRef{benjamini2000adaptive}{NetCoMi} \cr
  \insertRef{discordant2016}{NetCoMi} \cr
  \insertRef{farcomeni2007some}{NetCoMi} \cr
  \insertRef{fisher1992statistical}{NetCoMi} \cr
  \insertRef{gill2010statistical}{NetCoMi}
}
\seealso{
\code{\link{plot.diffnet}}
}
