% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netCompare_methods.R
\name{summary.microNetComp}
\alias{summary.microNetComp}
\alias{print.summary.microNetComp}
\title{Summary Method for Objects of Class netProps}
\usage{
\method{summary}{microNetComp}(
  object,
  groupNames = NULL,
  pAdjust = TRUE,
  showCentr = "all",
  numbNodes = 10L,
  digits = 3L,
  digitsPval = 6L,
  ...
)

\method{print}{summary.microNetComp}(x, ...)
}
\arguments{
\item{object}{object of class \code{netProps} inheriting from a call to
\code{\link{netCompare}}.}

\item{groupNames}{character vector with two elements giving the group names
corresponding to the two networks. If \code{NULL}, the names are adopted
from \code{object}}

\item{pAdjust}{logical. If \code{TRUE} the permutation p-values adjusted for
multiple testing are shown.}

\item{showCentr}{character vector indicating for which centrality measures
the results (values for both groups, difference and p-values resulting from
permutation tests) shall be printed. Possible values are "all", "degree",
"betweenness", "closeness", "eigenvector" and "none".}

\item{numbNodes}{integer indicating for how many nodes the centrality
values shall be printed. Defaults to 10 which means that the first 10 taxa
with highest absolute group difference of the specific centrality
measure are shown.}

\item{digits}{integer giving the number of decimal places to which the 
results are rounded. Defaults to 3L.}

\item{digitsPval}{integer giving the number of decimal places to which the 
p-values are rounded. Defaults to 6L.}

\item{...}{not used}

\item{x}{object of class \code{summary.microNetComp} inheriting from a call
to \code{\link{summary.microNetComp}}.}
}
\description{
The main results returned from \code{\link{netCompare}} are
  printed in a well-arranged format.
}
\seealso{
\code{\link{netCompare}}
}
