% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netAnalyze_methods.R
\name{summary.microNetProps}
\alias{summary.microNetProps}
\alias{print.summary.microNetProps}
\title{Summary Method for Objects of Class microNetProps}
\usage{
\method{summary}{microNetProps}(
  object,
  groupNames = NULL,
  showCentr = "all",
  numbNodes = NULL,
  digits = 5L,
  ...
)

\method{print}{summary.microNetProps}(x, ...)
}
\arguments{
\item{object}{object of class \code{microNetProps} inheriting from a call to
\code{\link{netAnalyze}}.}

\item{groupNames}{character vector with two elements giving the group names
corresponding to the two networks. If \code{NULL}, the names are adopted
from \code{object}. Ignored if \code{object} contains a single network.}

\item{showCentr}{character vector indicating for which centrality measures
the results shall be printed. Possible values are "all", "degree",
"betweenness", "closeness", "eigenvector" and "none".}

\item{numbNodes}{integer indicating for how many nodes the centrality
values shall be printed. Defaults to 10L for a single network and 5L for 
two networks. Thus, in case of a single network, the first 10 
nodes with highest centrality value of the specific centrality measure
are shown. If \code{object} contains two networks, for each centrality 
measure a splitted matrix is shown where the upper part contains the 
highest values of the first group, and the lower part the highest values of 
the second group.}

\item{digits}{integer giving the number of decimal places to which the 
results are rounded. Defaults to 5L.}

\item{...}{not used}

\item{x}{object of class \code{summary.microNetProps} inheriting from a call
to \code{\link{summary.microNetProps}}.}
}
\description{
The main results returned from \code{\link{netAnalyze}} are
  printed in a well-arranged format.
}
\seealso{
\code{\link{netConstruct}}, \code{\link{netAnalyze}}
}
